+++ modules/jaxb-xml-binding/geotk-xml-gml/src/main/java/org/geotoolkit/gml/xml/v311/DerivationUnitTermType.java
@@ -16,7 +16,6 @@
  */
 package org.geotoolkit.gml.xml.v311;
 
-import java.math.BigInteger;
 import javax.xml.bind.annotation.XmlAccessType;
 import javax.xml.bind.annotation.XmlAccessorType;
 import javax.xml.bind.annotation.XmlAttribute;
@@ -45,22 +44,20 @@ import javax.xml.bind.annotation.XmlType;
  */
 @XmlAccessorType(XmlAccessType.FIELD)
 @XmlType(name = "DerivationUnitTermType")
-public class DerivationUnitTermType
-    extends UnitOfMeasureType
-{
+public class DerivationUnitTermType extends UnitOfMeasureType {
 
     @XmlAttribute
-    protected BigInteger exponent;
+    protected Integer exponent;
 
     /**
      * Gets the value of the exponent property.
      * 
      * @return
      *     possible object is
-     *     {@link BigInteger }
+     *     {@link Integer }
      *     
      */
-    public BigInteger getExponent() {
+    public Integer getExponent() {
         return exponent;
     }
 
@@ -69,10 +66,10 @@ public class DerivationUnitTermType
      * 
      * @param value
      *     allowed object is
-     *     {@link BigInteger }
+     *     {@link Integer }
      *     
      */
-    public void setExponent(BigInteger value) {
+    public void setExponent(Integer value) {
         this.exponent = value;
     }
